Move benchmark utils to another file
